Add files via upload
Hummingbirds are key pollinators in many ecosystems, especially in flower gardens with large and colourful flowers. Flowers depend on pollinators like hummingbirds for reproduction while nectar from flowers are a great food source for hummingbirds. Thus, without hummingbirds, the flower population will die out and without flowers, hummingbirds will also die out. However, there is another factor that needs to be taken into account: predators that hunt these birds. If the population of predators is too large, the predators will hunt down all the hummingbirds. This would then lead to a decrease in the flower population. Hence, there needs to be a well-rounded balance between these three organisms in order for all species to survive and thrive in their environment. This automaton model has two modes: nature and chaotic. The “nature” mode shows the phenomenon described above. It displays how the hummingbirds and flowers have a symbiotic relationship as well as portray the end results of an unbalanced population of all species. In the chaotic mode, the opposite phenomenon occurs in which the hummingbirds become savages. They eat the predators and destroy the flowers. Due to the change in the animal hierarchy, the predators are going to mate and reproduce more rapidly as a survival instinct.